refactor: extract fitness analysis charts and scrub hook
